+++ src/playground_plug/playground_plug/modules/env/FFEnvState.hpp
@@ -13,7 +13,7 @@ struct FBModuleGraphComponentData;
 class FFEnvGUIState final
 {
   friend struct FFGUIState;
-  std::array<float, 1> graphKeyTimePct = {};
+  std::array<double, 1> graphKeyTimePct = {};
   friend std::unique_ptr<FBStaticModule> FFMakeEnvTopo();
 public:
   FB_NOCOPY_NOMOVE_DEFCTOR(FFEnvGUIState);
